Use IODetector2 as IODetector
